+++ packages/react/rollup.config.js
@@ -1,7 +1,7 @@
+import alias from "@rollup/plugin-alias";
 import babel from "@rollup/plugin-babel";
 import resolve from "@rollup/plugin-node-resolve";
 import postCssValues from "postcss-modules-values";
-import polyfill from "rollup-plugin-polyfill";
 import postcss from "rollup-plugin-postcss";
 import sourcemaps from "rollup-plugin-sourcemaps";
 import { terser } from "rollup-plugin-terser";
@@ -21,12 +21,17 @@ export default [
       },
     ],
     plugins: [
+      alias({
+        entries: {
+          "react-resize-detector": "react-resize-detector/build/withPolyfill",
+        },
+      }),
       resolve({
         extensions,
         browser: true,
         preferBuiltins: false,
         modulesOnly: true,
-        resolveOnly: [/^@amfa-team\/user-service.*$/],
+        resolveOnly: [/^@amfa-team\/room-service.*$/],
       }),
       sourcemaps(),
       postcss({
@@ -40,7 +45,6 @@ export default [
         extensions,
         plugins: [["@babel/plugin-transform-runtime", { useESModules: true }]],
       }),
-      polyfill(["webrtc-adapter", "abortcontroller-polyfill"]),
       ...extraPlugins,
     ],
   },
@@ -54,12 +58,17 @@ export default [
       },
     ],
     plugins: [
+      alias({
+        entries: {
+          "react-resize-detector": "react-resize-detector/build/withPolyfill",
+        },
+      }),
       resolve({
         extensions,
         browser: true,
         preferBuiltins: false,
         modulesOnly: true,
-        resolveOnly: [/^@amfa-team\/user-service.*$/],
+        resolveOnly: [/^@amfa-team\/room-service.*$/],
       }),
       sourcemaps(),
       postcss({
@@ -73,7 +82,6 @@ export default [
         extensions,
         plugins: [["@babel/plugin-transform-runtime", { useESModules: false }]],
       }),
-      polyfill(["webrtc-adapter", "abortcontroller-polyfill"]),
       ...extraPlugins,
     ],
   },